pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "wlanpi-fpms2"
version = "0.1.0"
description = "WLANPi Front Panel Menu System 2 — decoupled state/display architecture"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
    "httpx>=0.27",
    "pydantic>=2.7",
    "websockets>=12.0",
    "textual>=0.61",
    "Pillow>=10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.23",
    "respx>=0.21",
    "httpx>=0.27",
]

[project.scripts]
wlanpi-fpms2 = "wlanpi_fpms2.__main__:main"
wlanpi-fpms2-screen = "wlanpi_fpms2.interfaces.screen.client:main"
wlanpi-fpms2-tui = "wlanpi_fpms2.interfaces.tui.app:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["wlanpi_fpms2*"]

[tool.setuptools.package-data]
"wlanpi_fpms2.interfaces.screen" = ["fonts/*", "images/128x128/*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
